+++ release_notes.md
@@ -263,4 +263,5 @@ BugFixes without ticket
 
 ## BugFixes
 1. [TH2-5222] - Fix `DownloadMessagesByPageGzip`, it constructed `DownloadMessagesByPageByGroupsGzip` incorrectly.
-2. [TH2-5243] - Fix backward compatibility issues with `GetMessagesByBookByGroups` and `GetMessagesByPageByGroups`.
+2. [TH2-5243] - Fix backward compatibility issues with `GetMessagesByBookByGroups` and `GetMessagesByPageByGroups`,
+   added 'limit' and 'search_direction' parameters to 'post_download_messages'.